+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Simplified duckdb conn kwargs typing
+
 ## [2026.2.0] - 2026-02-07
 
 ### Added